Clean up + setting up docker build

+++ blockchain-contract/dnsrecord/inet-dns-record.sol
@@ -1,13 +1,13 @@
 pragma solidity 0.6.3;
 
 contract InetDnsRecord {
-    mapping (string => mapping(int => string)) internal _dnsMapping;
+    mapping (string => mapping(uint16 => string)) internal _dnsMapping;
 
-    function addRecord(string memory key, int8 recType, string memory recValue) public {
+    function addRecord(string memory key, uint16 recType, string memory recValue) public {
         _dnsMapping[key][recType] = recValue;
     }
 
-    function getRecord(string memory key, int8 recType) public view returns (string memory){        
+    function getRecord(string memory key, uint16 recType) public view returns (string memory){        
         return _dnsMapping[key][recType];
     }
 }
